requirements.txt
django
python-decouple
supabase
razorpay
setuptools
requests